// frontend/src/types/api.generated.ts
export interface components {
  schemas: {
// ...
    SubscribeRequest: {
      /**
       * Format: email
       * @description Email адрес для подписки
       */
      email: string;
      pdp_consent: boolean;
      marketing_consent: boolean;
      pdp_consent_text_version: string;
      marketing_consent_text_version: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}